@@ -3,7 +3,6 @@ package mounter
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"strings"
@@ -23,8 +22,9 @@ type Mounter interface {
 }
 
 const (
-	s3fsMounterType   = "s3fs"
-	rcloneMounterType = "rclone"
+	S3fsMounterType   = "s3fs"
+	RcloneMounterType = "rclone"
+	GoofysMounterType = "goofys"
 )
 
 func NewMounter(metadata *s3.Metadata, config *s3.Config) (Mounter, error) {
@@ -33,25 +33,29 @@ func NewMounter(metadata *s3.Metadata, config *s3.Config) (Mounter, error) {
 		mounter = config.Mounter
 	}
 	switch mounter {
-		case s3fsMounterType:
-			return newS3fsMounter(metadata, config)
-		case rcloneMounterType:
-			return newRcloneMounter(metadata, config)
-		default:
-			return newS3fsMounter(metadata, config)
+	case S3fsMounterType:
+		return newS3fsMounter(metadata, config)
+	case RcloneMounterType:
+		return newRcloneMounter(metadata, config)
+	case GoofysMounterType:
+		return newGoofysMounter(metadata, config)
+	default:
+		klog.Errorf("unknown mounter %s, using default mounter %s", mounter, S3fsMounterType)
+		return newS3fsMounter(metadata, config)
 	}
 }
 
-func fuseMounter(path string, command string, args []string) error {
+func fuseMount(path string, command string, args []string) error {
 	cmd := exec.Command(command, args...)
-	klog.Infof("Mounting fuse with command: %s and args: %s", command, args)
+	klog.Infof("Mount fuse with command: %s with args %s", command, args)
+
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("Error fuseMount command: %s\nargs: %s\noutput", command, args)
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		return fmt.Errorf("Mount fuse mount with command: %s with args %s\nerror: %s", command, args, string(out))
 	}
-
 	return waitForMount(path, 10*time.Second)
 }
 
@@ -88,7 +92,7 @@ func FuseUmount(path string) error {
 		return nil
 	}
 	klog.Infof("Found fuse PID %v of fuse mount %s, checking if it still runs", process.Pid, path)
-	return waitForProcess(process)
+	return waitForProcess(process, 0)
 }
 
 func findFuseMountProcess(path string) (*os.Process, error) {
@@ -112,17 +116,19 @@ func findFuseMountProcess(path string) (*os.Process, error) {
 
 func getCmdline(pid int) (string, error) {
 	cmdlineFile := fmt.Sprintf("/proc/%v/cmdline", pid)
-	cmdline, err := ioutil.ReadFile(cmdlineFile)
+	cmdline, err := os.ReadFile(cmdlineFile)
 	if err != nil {
 		return "", err
 	}
 	return string(cmdline), nil
 }
 
-func waitForProcess(process *os.Process) error {
-	var elapsed time.Duration
+func waitForProcess(process *os.Process, backoff int) error {
+	if backoff == 20 {
+		return fmt.Errorf("time waiting for PID %v to end", process.Pid)
+	}
+
 	interval := 2 * time.Second
-	timeout := 120 * time.Second
 	for {
 		cmdline, err := getCmdline(process.Pid)
 		if err != nil {
@@ -134,15 +140,27 @@ func waitForProcess(process *os.Process) error {
 			proc, _ := os.FindProcess(process.Pid)
 			return proc.Kill()
 		}
-		if err := process.Signal(syscall.Signal(0)); err != nil {
+		if err = process.Signal(syscall.Signal(0)); err != nil {
 			klog.Warningf("Fuse process does not seem active or we are unprivileged: %s", err)
 			return nil
 		}
 		klog.Infof("Fuse process with PID %v still active, waiting...", process.Pid)
-		time.Sleep(interval)
-		elapsed = elapsed + interval
-		if elapsed >= timeout {
-			return fmt.Errorf("timeout waiting for PID %v to end", process.Pid)
-		}
+		time.Sleep(interval + time.Duration(backoff*100))
+		return waitForProcess(process, backoff+1)
 	}
+}
+
+func CheckMount(volumeId, path string) (bool, error) {
+	isMount, err := mount.New("").IsLikelyNotMountPoint(path)
+	if err != nil {
+		if os.IsNotExist(err) {
+			if err := os.MkdirAll(path, 0750); err != nil {
+				return false, err
+			}
+		}
+		isMount = false
+	} else {
+		return false, err
+	}
+	return isMount, nil
 }
