@@ -1,10 +1,6 @@
 package constant
 
 const (
-	DriverName    = "io.github.leryn.csi.s3driver"
-	VendorVersion = "v0.1.0"
-
-	TypeKey      = "mounter"
-	BucketKey    = "bucket"
-	VolumePrefix = "prefix"
+	TypeKey   = "mounter"
+	BucketKey = "bucket"
 )
