@@ -2,11 +2,8 @@ package mounter
 
 import (
 	"fmt"
-	"os"
-	"os/exec"
-
 	"github.com/leryn1122/csi-s3/pkg/s3"
-	"k8s.io/klog/v2"
+	"os"
 )
 
 const s3fsCmd = "s3fs"
@@ -41,7 +38,7 @@ func (s3fs *s3fsMounter) Mount(_ string, target string) error {
 		return err
 	}
 	args := []string{
-		fmt.Sprintf("%s:/%s", s3fs.metadata.BucketName, s3fs.metadata.FsPath),
+		fmt.Sprintf("%s:/%s", s3fs.metadata.BucketName, s3fs.metadata.FsPathPrefix),
 		target,
 		"-o", "use_path_request_style",
 		"-o", fmt.Sprintf("url=%s", s3fs.url),
@@ -49,7 +46,7 @@ func (s3fs *s3fsMounter) Mount(_ string, target string) error {
 		"-o", "allow_other",
 		"-o", "mp_umask=000",
 	}
-	return fuseMount(s3fsCmd, args)
+	return fuseMount(target, s3fsCmd, args)
 }
 
 func writeS3fsPassword(pwFileContent string) error {
@@ -64,17 +61,6 @@ func writeS3fsPassword(pwFileContent string) error {
 	}
 	if pwFile.Close() != nil {
 		return err
-	}
-	return nil
-}
-
-func fuseMount(command string, args []string) error {
-	cmd := exec.Command(command, args...)
-	klog.Infof("Mount fuse with command: %s with args %s", command, args)
-
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("Mount fuse with command: %s with args %s\nerror: %s", command, args, string(out))
 	}
 	return nil
 }
